adding the parser
In order to get the information about the exons I need to get it from
the RefSeq web site.
Since the information about the exons in the Refseq web site is hidden I
can't get the information I need from the Refseq source code. Luckily, I
found out that the Refseq's source code contains a link to the page that
shows the information about the exons. I could find the infix to the
above link by searching "/nuccore" in the Refseq's source code. In this
method I'm extracting the infix of the "nuccore" URL and returning the
full URL as an output.
The parser also contain methods that read the source code of a web site
and save it in a file, read from files etc.